+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - Selecting a search history item used to reset params that were not part of the search
 - Trend diagram legend missing
+- Internal search links in sidebar does not activate relevant search/result tabs [#450](https://github.com/spraakbanken/korp-frontend/issues/450)
 - Comparison result not showing if a search is not done first [#413](https://github.com/spraakbanken/korp-frontend/issues/413)
 - Extended search: do not cache operator options across corpora [#409](https://github.com/spraakbanken/korp-frontend/issues/409)
 - Error when logging out while protected corpora are selected [#440](https://github.com/spraakbanken/korp-frontend/issues/440)